requirements.txt
PySide6
sqlcipher3-wheels
password-strength
favicon
requests